entry.size is now uint16_t instead of uint8_t

+++ src/object.cpp
@@ -68,7 +68,7 @@ bool Object::isSubValid(uint8_t subindex)
     return subindex < subNumber;
 }
 
-uint8_t Object::getSize(uint8_t subindex)
+uint16_t Object::getSize(uint8_t subindex)
 {
     if (!isSubValid(subindex))
         return 0;
